rename domain ReactChilds to ReactChildren
