+++ components/Paragraph/view.tsx
@@ -1,10 +1,10 @@
 import React, { PureComponent } from 'react'
-import TypografReactNode from '~/domain/typografy/reactChilds'
+import TypografReactNode from '~/domain/typografy/reactChildren'
 
 class Paragraph extends PureComponent {
   render() {
     const { children } = this.props
-    return TypografReactNode.process(<p>{children}</p>)
+    return TypografReactNode.processElement(<p>{children}</p>)
   }
 }
